refactor to have newphotoform component
